Migrate standard answers implementations to ee/
